+++ src/docs/changelog.md
@@ -4,6 +4,8 @@ Changelog                           {#changelog}
 0.22.0
 ------
 
+- created sprite animation example
+  ![sprite animation](/images/SpriteAnimation.gif)
 - ownership of widgets transfers to the engine the moment they are children of a dialog
 - added mesh to world bake functionality
 - removed all using namesapce std;